chore(build tasks)
was not properly importing a module for the development task. It is now
running properly, work that still needs to be done however is
abstracting the paths so that they are defined in the config.

also went through the build tasks and refactored out all the variable
declarations to be be a bit cleaner.